third_party/blink/web_tests/external/wpt/css/css-scroll-snap/snap-to-combination-of-two-elements-2-expected.txt

This is a testharness.js-based test.
[FAIL] Snap to points of combinations of two different elements
  assert_equals: expected 50 but got 200
Harness: the test ran to completion.
